Move OrderedStringTestGenerator into CipherStash::Client
